MT19937 cipher
You can create a trivial stream cipher out of any PRNG;
use it to generate a sequence of 8 bit outputs and call those
outputs a keystream. XOR each byte of plaintext with each
successive byte of keystream.

Write the function that does this for MT19937 using a 16-bit seed.
Verify that you can encrypt and decrypt properly. This code should
look similar to your CTR code.